requirements.txt
Django==4.2.17
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
pillow==11.1.0
django-filter==25.1
djangorestframework_simplejwt==5.5.0
